refactor: extract stream opener to opener.rs (Phase 4)
Move Group E (StreamOpener + OpenStreams impl + stream_open_conn_error +
classify_start_outcome) and the stream_open_identity + downcall_clamp test
modules out of lib.rs into opener.rs. StreamOpener re-exported with pub use
(public path preserved); internals cfg(test)-gated pub(crate) use. Pure move;
no behavior/API change. 176 tests / 2 ignored.
